Move `packages/ov-attribution` to `py_packages/openverse-attribution`
