+++ .github/filters.yml
@@ -9,7 +9,7 @@ catalog:
   - docker-compose.yml
 api:
   - api/**
-  - packages/ov-attribution/**
+  - py_packages/**
   - docker/db/**
   - docker/es/**
   - docker/nginx/**
