+++ api/pyproject.toml
@@ -36,7 +36,7 @@ dependencies = [
   "python-xmp-toolkit >=2.0.2, <3",
   "sentry-sdk >=1.40.6, <2",
   "uvicorn[standard] >=0.27.1, <0.28",
-  "ov-attribution @ file:///${PROJECT_ROOT}/../packages/ov-attribution"
+  "openverse-attribution @ file:///${PROJECT_ROOT}/../py_packages/openverse-attribution"
 ]
 
 [tool.pdm]
@@ -45,7 +45,7 @@ distribution = false
 [tool.pdm.dev-dependencies]
 # These packages override their main counterparts with an editable installation.
 overrides = [
-  "-e ov-attribution @ file:///${PROJECT_ROOT}/../packages/ov-attribution"
+  "-e openverse-attribution @ file:///${PROJECT_ROOT}/../py_packages/openverse-attribution"
 ]
 dev = [
   "ipython >=8.22.1, <9",
