+++ automations/python/workflows/set_matrix_images.py
@@ -29,12 +29,16 @@ includes = {
     },
     "catalog": {"image": "catalog", "target": "cat"},
     "ingestion_server": {"image": "ingestion_server", "target": "ing"},
-    "api": {"image": "api", "target": "api", "build-contexts": "packages=./packages"},
+    "api": {
+        "image": "api",
+        "target": "api",
+        "build-contexts": "packages=./py_packages",
+    },
     "api_nginx": {
         "image": "api_nginx",
         "context": "api",
         "target": "nginx",
-        "build-contexts": "packages=./packages",
+        "build-contexts": "packages=./py_packages",
     },
     "frontend": {"image": "frontend", "target": "app", "build-contexts": "repo_root=."},
     "frontend_nginx": {
