+++ py_packages/openverse-attribution/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "ov-attribution"
+name = "openverse-attribution"
 version = "0.1.0"
 description = "Provides utilities related to open licenses, tools, and attribution."
 readme = "README.md"
